// src/app/model/trabajador.ts
export class  Trabajador {

    id?: number;
    rut: string;
    //digito_verificador :int;
    nombre: string;
    ap_paterno: string;
    ap_materno: string;
    //direccion: string;
    id_sector: number;
    fecha_ingreso: string;
    fecha_baja?:Date;
    telefono: number;
    email: string;
   
    }